Fix connection testing logic
Simplify it by forwarding exceptions in `refresh_connection()` and drop
`verify_access()` entirely. In the original `deadline-cloud-plugin`
code, those 2 methods were not used consistantly: either
`verify_access()` was used alone or it was following the
`refresh_connection()` method. It was also odd that refresh_connection()
was called *inside* verify_access() .

Now we only rely on `refresh_connection()` because it has both benefits
of refreshing the connection and logging an error + raising an exception
if something's wrong with the connection.