Change defaults for proxy_crt
Today we do not handle the case where a certificate file is provided
(because certificates are in fact being handled directly in the Qarnot
profile/payload). Hence we won't use a file path has a default for
`proxy_crt` to avoid confusion. For now, we will use a more sensible
default (we use "True" but it can be anything, as long as it is not empty
because that's what the payload expects)