Prevent having a 'b' prefix in DEADLINE_CRT
The Deadline certificate is converted to base64 which returns bytes.
And, when casting bytes to string in Python 3, the content is prefixed
with 'b'. This behaviour does not happen with Python 2 (used in
Deadline) so the problem can only be noticed when testing the
qarnot_render_deadline module with Python 3 during development.
In any case, instead of casting to string, the base64 certificate is now
decoded as ASCII, which also returns a string in Python 3, but without
being prefixed with 'b'.